Kristofer Sandlund patch
We have changed some parts of the writers to better fit with the "MATLAB 7 MAT-File Format" document. The current file format output by JMatIO can be read by normal MATLAB, but it has some incompatible differences with the file format specification and the files cannot be read by Octave.
Since it seems useful to have compatiblity with Octave as well, we changed  the datatype of some fields. We added the short format for the OSArrayTag (when size is 4 or less) and changed the matrix name field from UTF-8 to int8 since that is what is stated in the specification

In addition, we have added support for INT16 through new class MLInt16 and support for it to the writers and the reader.



git-svn-id: http://svn.code.sf.net/p/jmatio/code/trunk@87 70b73bdd-3b7a-4af7-8b34-4a2dea44f45c